rename utils `crf_name` ~> `crf` and update refs
