+++ bridge/callbacks/save.py
@@ -6,7 +6,7 @@ import dash
 import pandas as pd
 from dash import dcc, Input, Output, State
 
-from bridge.utils.crf_name import get_crf_name
+from bridge.utils.crf import get_crf_name
 from bridge.utils.trigger_id import get_trigger_id
 
 pd.options.mode.copy_on_write = True
